% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsum.R
\name{calc_max_edges}
\alias{calc_max_edges}
\title{Calculate maximum possible edges}
\usage{
calc_max_edges(rs, cs)
}
\arguments{
\item{rs}{numeric vector, row sums}

\item{cs}{numeric vector, col sums}
}
\value{
maximum possible number of edges
}
\description{
For a set of row and column sums compute maximum number of
edges assuming no self-loops. Any row/column sum that is zero implies that
node has no outgoing/incoming edges.
}
\author{
Douglas Ashton
}

