% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maxEntropy.R
\name{maxEntropy}
\alias{maxEntropy}
\title{Find the so-called maximum entropy solution}
\usage{
maxEntropy(rs, cs, minError = 1e-18)
}
\arguments{
\item{rs}{NumericVector the row sums of the matrix}

\item{cs}{NumericVector the column sums of the matrix}
}
\value{
A matrix that satisfies the row and column sum constraints or FALSE
 if the algorithm failed to converge
}
\description{
The maximum entropy solution is obtained by spreading the edge
weights as evenly as possible throught the weighted adjacency matrix. This
version forces the diagonal elements to be zero along with any rows (columns)
that have a zero row (column) sum in the input.

Missing values are allowed but will result in a warning. The usefulness of
such results is up to the user.
}

