% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-neast.R
\docType{data}
\name{neast}
\alias{neast}
\title{North East Bank data}
\format{A \code{\link{data.frame}} with 6 observations.}
\usage{
neast
}
\description{
A demo balance sheet for six fictional banks
}
\examples{
neast
}

