% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-neast.R
\name{neastTrue}
\alias{neastTrue}
\title{North East Bank Full Network data}
\format{A 6x6 matrix}
\usage{
neastTrue
}
\description{
The full network from which the neast balance sheet is generated. 
This is useful for testing how well networks generated with unsum match the
starting network.
}
\examples{
neastTrue
}

