% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netted_matrix.R
\name{netted_matrix}
\alias{netted_matrix}
\alias{netted_matrix.default}
\alias{netted_matrix.unsumnet}
\title{Calculate the netted matrix}
\usage{
netted_matrix(x)

\method{netted_matrix}{default}(x)

\method{netted_matrix}{unsumnet}(x)
}
\arguments{
\item{x}{A square numeric matrix or an object of class \code{unsumnet}}
}
\value{
A numeric matrix with the positive elements of x-x' (or x$AW - x$AW'
for \code{unsumnet} object) and zero for negative elements.
}
\description{
Subtract the transpose and set negative elements to zero for a weighted
adjacency matrix.
}
\examples{
netted_matrix(matrix(rnorm(25), nrow=5))
}

