% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nettedMatrix.R
\name{nettedMatrix}
\alias{nettedMatrix}
\title{Calculate the netted matrix}
\usage{
nettedMatrix(w)
}
\arguments{
\item{w}{A square numeric matrix}
}
\value{
A numeric matrix with the positive elements of w-w'
}
\description{
Subtract the transpose and keep positive elements of a weighted
adjacency matrix.
}

