% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nettedMatrix.R
\name{nettedMatrix.unsumnet}
\alias{nettedMatrix.unsumnet}
\title{Calculate the netted matrix (default)}
\usage{
\method{nettedMatrix}{unsumnet}(x)
}
\arguments{
\item{x}{An \code{unsumnet} object usually from \code{\link{unsum}}.}
}
\value{
A numeric matrix with the positive elements of aw-aw'
}
\description{
Subtract the transpose and keep positive elements of a weighted
adjacency matrix.
}

