% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.unsumnet}
\alias{plot.unsumnet}
\title{Plot method for an \code{unsumnet} network}
\usage{
\method{plot}{unsumnet}(x, ...)
}
\arguments{
\item{x}{an \code{unsumnet} object}

\item{...}{extra options passed through to \code{\link{plot.igraph}}}
}
\description{
Passes the weighted adjacency matrix to \code{\link{plotUnsum}}
which is a wrapper around the igraph plot. You will need to install
\code{igraph} to use this plot method.
}

