% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_unsum}
\alias{plot_unsum}
\title{Plot function for a weighted adjacency matrix}
\usage{
plot_unsum(aw, ...)
}
\arguments{
\item{aw}{a matrix}

\item{...}{arguments passed through to \code{plot.igraph}}
}
\description{
This will create a directed \code{igraph} network and call its
plotting methods
}
\examples{
plot_unsum(rbind(c(0,1,2),c(0,0,1), c(1,0,0)))
}

