% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotUnsum}
\alias{plotUnsum}
\title{Plot function for a weighted adjacency matrix}
\usage{
plotUnsum(aw, ...)
}
\arguments{
\item{aw}{a matrix}

\item{...}{}
}
\description{
This will create a directed \code{igraph} network and call its
plotting methods
}
\examples{
plotUnsum(rbind(c(0,1,2),c(0,0,1), c(1,0,0))
}

