% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processInput.R
\name{processInput}
\alias{processInput}
\title{Process the input constraints}
\usage{
processInput(constraints)
}
\arguments{
\item{constraints}{A matrix or data frame containing the row and column sums}
}
\value{
an object ready for sending to the C++ functions as input
}
\description{
This function checks and prepares the network constraints before
calling the C++ code.
}
\author{
Douglas Ashton
}

