% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.unsumnet}
\alias{summary.unsumnet}
\title{Summarising Unsummed Networks}
\usage{
\method{summary}{unsumnet}(x)
}
\arguments{
\item{x}{an object of class "unsumnet", usually from a call to
\code{\link{unsum}}.}
}
\description{
\code{\link{summary}} method for class "unsumnet"
}

