% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.unsumnet}
\alias{print.unsumnet}
\title{Print method for \code{unsumnet}}
\usage{
\method{print}{unsumnet}(x, ...)
}
\arguments{
\item{x}{an object, usually generated from \code{\link{unsum}}.}

\item{...}{passed on to other print methods.}
}
\description{
Print method for \code{unsumnet}
}

